requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
pillow
opencv-python-headless
numpy
pymupdf
pdfplumber
pytesseract
openai
httpx
rapidfuzz
validators
pytest
pytest-asyncio
requests
